Implement TryFrom traits to sign and encode entries
