+++ p2panda-rs/src/atomic/mod.rs
@@ -2,7 +2,7 @@ use crate::Result;
 
 mod author;
 mod entry;
-mod entry_encoded;
+mod entry_signed;
 mod hash;
 mod log_id;
 mod message;
@@ -11,7 +11,7 @@ mod seq_num;
 
 pub use author::Author;
 pub use entry::Entry;
-pub use entry_encoded::EntryEncoded;
+pub use entry_signed::EntrySigned;
 pub use hash::Hash;
 pub use log_id::LogId;
 pub use message::{Message, MessageAction, MessageFields, MessageValue, MessageVersion};
@@ -22,7 +22,7 @@ pub use seq_num::SeqNum;
 pub mod error {
     pub use super::author::AuthorError;
     pub use super::entry::EntryError;
-    pub use super::entry_encoded::EntryEncodedError;
+    pub use super::entry_signed::EntrySignedError;
     pub use super::hash::HashError;
     pub use super::message::{MessageError, MessageFieldsError};
     pub use super::message_encoded::MessageEncodedError;
